chore(cleanup): modularizes code
