+++ src/components/options/options.actions.js
@@ -1,0 +1,3 @@
+export const ADD_CHARACTER = 'add_character'
+export const LOAD_CHARACTERS = 'load_characters'
+export const SAVE_CHARACTERS = 'save_characters'
