+++ src/components/options/options.scss
@@ -1,0 +1,15 @@
+.dicebot-options {
+  pre {
+    background-color: lightgray;
+  }
+
+  &__character {
+    border: 3px solid black;
+    padding: 5px;
+    margin: 5px;
+
+    &--input {
+      margin-bottom: 5px;
+    }
+  }
+}
