+++ src/index.html
@@ -1,4 +1,7 @@
 <!DOCTYPE html>
+<head>
+    <meta charset="UTF-8">
+</head>
 <html>
     <body>
         <div id="root"></div>
