--- src/index.scss
@@ -1,15 +1,0 @@
-.dicebot-options {
-    pre {
-      background-color: lightgray;
-    }
-
-    &__character {
-      border: 3px solid black;
-      padding: 5px;
-      margin: 5px;
-
-      &--input {
-        margin-bottom: 5px;
-      }
-    }
-}
